backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.99.0
anyio==4.13.0
asyncpg==0.31.0
certifi==2026.4.22
click==8.3.3
distro==1.9.0
docstring_parser==0.18.0
fastapi==0.136.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.13
importlib_metadata==8.7.1
jiter==0.14.0
openai==2.34.0
opentelemetry-api==1.41.1
opentelemetry-sdk==1.41.1
opentelemetry-semantic-conventions==0.62b1
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
python-dotenv==1.2.2
PyYAML==6.0.3
redis==7.4.0
sniffio==1.3.1
starlette==1.0.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
zipp==3.23.1
